:truck: Moved models to DB directory
