+++ src/dao/DBBasedManagers/CartManager/CartManagerDBBased.js
@@ -1,5 +1,5 @@
 import { Cart } from "../../../main/Cart/Cart.js";
-import cartModel from "../../models/cart.model.js";
+import cartModel from "../../DBBasedManagers/models/cart.model.js";
 import { mongoose } from "mongoose";
 import { productManager } from "../ManagerSystem/ManagerSystem.js";
