+++ src/dao/DBBasedManagers/MessageManager/MessageManagerDBBased.js
@@ -1,5 +1,5 @@
 import { Message } from "../../../main/Message/Message.js";
-import messageModel from "../../models/message.model.js";
+import messageModel from "../../DBBasedManagers/models/message.model.js";
 
 export class MessageManagerDBBased {
   assertSatisfiesAllRequiredParameters = ({
